Change file resource to exec resource.
Found out that the file resource type does not have the attribute
'onlyif'. This only works for exec resource type.

Therefore, rewrote that section to execute 'mkdir'.

Also realised the flag used in the test for onlyif is incorrect. '-d'
test if if file IS a directory, whereas '-e' tests if file (or directory) EXISTS.